% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrobiosum.R
\docType{data}
\name{agrobiosum}
\alias{agrobiosum}
\title{Yearwise Yield data of the Agrodiversity experiment}
\format{
A dataframe with 92 rows and 9 variables:
\describe{
\item{SITE}{Site ID number}
\item{Country}{Country}
\item{Year}{Year}
\item{G1_YA}{Average harvest Dry Matter Yield per year of fast-establishing grass}
\item{G2_YA}{Average harvest Dry Matter Yield per year of persistent grass}
\item{L1_YA}{Average harvest Dry Matter Yield per year of fast-establishing legume }
\item{L2_YA}{Average harvest Dry Matter Yield per year of persistent legume }
\item{WEED_YA}{Average harvest Dry Matter Yield per year of weed species }
\item{HARV_YA}{Average harvest Dry Matter Yield per year}
}
}
\source{
\url{https://repositori.udl.cat/handle/10459.1/60560}
}
\usage{
agrobiosum
}
\description{
agrobiosum dataset provides information about Species-specific biomass measurements (yield per species and of weeds) at 33 sites summarized by year.
}
\keyword{datasets}
