% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroclimatedoc.R
\docType{data}
\name{agroclimate}
\alias{agroclimate}
\title{Climate data of the Agrodiversity experiment}
\format{
A dataframe with 39,712 rows and 9 variables:
\describe{
\item{Site}{Site ID number}
\item{Day}{Day of data collected}
\item{Month}{Month of data collected}
\item{Year}{Year of data collected}
\item{Date}{Date}
\item{Precip}{Daily precipitation (mm/ay)}
\item{air_min}{Minimum daily air temperature in degree celcius}
\item{air_mean}{Mean daily air temperature in degree celcius}
\item{air_max}{Maximum daily air temperature in degree celcius}
}
}
\source{
\url{https://repositori.udl.cat/handle/10459.1/60560}
}
\usage{
agroclimate
}
\description{
The \code{agroclimate} dataset has information about  daily climate data for each site for the duration of the experiment.
The climate data series begins on the sowing date of the experiment at each site.
The variables include daily temperatures and precipitation.
}
\section{References}{


Kirwan, L., Lüscher, A., Sebastià, M.T., Finn, J.A., Collins, R.P., Porqueddu, C., Helgadottir, A., Baadshaug, O.H., Brophy, C., Coran, C. and Dalmannsdottir, S., 2007. Evenness drives consistent diversity effects in intensive grassland systems across 28 European sites. Journal of Ecology, 95(3), pp.530-539.

Kirwan, L., Connolly, J., Brophy, C., Baadshaug, O.H., Belanger, G., Black, A., Carnus, T., Collins, R.P., Čop, J., Delgado, I. and De Vliegher, A., 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 2014, vol. 95, num. 9, p. 2680-2680.
}

\keyword{datasets}
