% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agrodiversitydata-package.R
\docType{package}
\name{Agrodiversitydata}
\alias{Agrodiversitydata}
\title{Agrodiversitydata:  Datasets of agrodiversity experiment.}
\description{
Agrodiversity experiment: A coordinated, continental-scale 33 site experiment that investigated plant diversity and ecosystem function in intensively-managed grassland communities.

#'
}
\section{Details}{


The package Agrodiversitydata provides six datasets:  Useful for ecologists studying relationships between diversity and ecosystem function and agronomists interested in sustainable intensification.
\enumerate{
\item \code{\link{agrosite}}: Information about individual sites in the experiment specifying details of country, species, treatment, harvest etc. .
\item \code{\link{agroclimate}}: Information about daily temperature and precipitation for each site.
\item \code{\link{agrosoil}}: The soil characteristics and concentration of soil nutrients for each site.
\item \code{\link{agrobiomass}}: The species-specific biomass measurements (yield per species and of weeds) for all harvests for up to four years at 33 sites.
\item \code{\link{agroforage}}: Details about forage quality.
\item \code{\link{agrobiosum}}: Species-specific biomass measurements (yield per species and of weeds) at 33 sites summarized by year.
}
}

\section{Examples}{


To load agrobiomass data

data(agrobiomass)
}

\section{References}{

Kirwan, L., Lüscher, A., Sebastià, M.T., Finn, J.A., Collins, R.P., Porqueddu, C., Helgadottir, A., Baadshaug, O.H., Brophy, C., Coran, C. and Dalmannsdottir, S., 2007. Evenness drives consistent diversity effects in intensive grassland systems across 28 European sites. Journal of Ecology, 95(3), pp.530-539.

Kirwan, L., Connolly, J., Brophy, C., Baadshaug, O.H., Belanger, G., Black, A., Carnus, T., Collins, R.P., Čop, J., Delgado, I. and De Vliegher, A., 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 2014, vol. 95, num. 9, p. 2680-2680.
}

\seealso{
For More information about data :\url{https://t-stor.teagasc.ie/handle/11019/774}
}
