% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrosoil.R
\docType{data}
\name{agrosoil}
\alias{agrosoil}
\title{Soil data of the Agrodiversity experiment}
\format{
A dataframe with 32 rows and 17 variables:
\describe{
\item{SITE}{Site ID number}
\item{Carbonates}{Soil Carbonates}
\item{EC}{Soil electrical conductivity (ds m-1)}
\item{Silt}{Percent silt content in soil}
\item{Silt_fine}{Percent fine silt content in soil}
\item{Clay}{Percent clay content in soil}
\item{Sand}{Percent sand content in soil}
\item{OM}{Percent organic matter}
\item{Soil_Type}{Soil type}
\item{Humidity}{Percent humidity}
\item{Ca}{Calcium concentration (ppm)}
\item{K}{Potassium concentration (ppm)}
\item{N-NO3}{Nitrate concentration (ppm)}
\item{Mg}{Magnesium concentration (ppm)}
\item{Na}{Sodium concentration (ppm)}
\item{P}{Phosphorus concentration (ppm)}
\item{pH}{Soil pH}
}
}
\source{
\url{https://repositori.udl.cat/handle/10459.1/60560}
}
\usage{
agrosoil
}
\description{
agrosoil dataset provides information about soil characteristics and concentration of soil nutrients for each site.
}
\keyword{datasets}
