% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass1.R
\docType{data}
\name{Biomass1}
\alias{Biomass1}
\title{Biomass1 data of the Agrodiversity experiment}
\format{
A dataframe with 15372 rows and 22 variables:
\describe{
\item{SITE}{Site ID number}
\item{COUNTRY}{Country}
\item{YEAR}{Year}
\item{YEARN}{Experimental year number}
\item{NH}{Number of harvests – number of times the whole plot was cut in a year}
\item{HARVEST}{Harvest number (within year)}
\item{HARVEST DATE}{Date of harvest}
\item{PLOT}{Plot number as per design (1–30 = core design; 31–48 = treatment plots; 49–68 = additional plots at sites 45 and 46)}
\item{TREAT}{Indicator variable: 1=basic 30 plots; 2 and 3=additional treatment plots (some sites implemented two levels of additional treatments)}
\item{REP}{Replicate number (applies only to sites 15, 45, and 46)}
\item{G1}{Initial sown proportion of fast-establishing grass}
\item{G2}{Initial sown proportion of persistent grass}
\item{L1}{Initial sown proportion of fast-establishing legume}
\item{L2}{Initial sown proportion of persistent legume}
\item{E}{Initial sown evenness}
\item{DENS}{Indicator variable: high=high level of initial sown biomass, low = low level (60\%of high)}
\item{G1_Y}{Harvest Dry Matter Yield of fast-establishing grass (t/ha)}
\item{G2_Y}{Harvest Dry Matter Yield of persistent grass (t/ha)}
\item{L1_Y}{Harvest Dry Matter Yield of fast-establishing legume (t/ha)}
\item{L2_Y}{Harvest Dry Matter Yield of persistent legume (t/ha)}
\item{WEED_Y}{Harvest Dry Matter Yield of weed species (t/ha)}
\item{HARV_YIELD}{Total Harvest Dry Matter Yield (t/ha)}
}
}
\source{
Laura Kirwan, John Connolly, Caroline Brophy, Ole Baadshaug, Gilles Belanger, Alistair Black, Tim Carnus, Rosemary Collins, Jure Čop, Ignacio Delgado, Alex De Vliegher, Anjo Elgersma, Bodil Frankow-Lindberg, Piotr Golinski, Philippe Grieu, Anne-Maj Gustavsson, Áslaug Helgadóttir, Mats Höglind, Olivier Huguenin-Elie, Marit Jørgensen, Žydrė Kadžiulienė, Tor Lunnan, Andreas Lüscher, Päivi Kurki, Claudio Porqueddu, M.-Teresa Sebastia, Ulrich Thumm, David Walmsley, John Finn. 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 95(9), pp. 2680-2680.

Data available at :\url{https://doi.org/10.6084/m9.figshare.3560352.v1}
}
\usage{
Biomass1
}
\description{
The \code{Biomass1} dataset provides information about the species-specific biomass measurements (yield per species and of weeds) for all harvests for up to four years at 33 sites.
}
\keyword{datasets}
