% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\docType{data}
\name{Climate}
\alias{Climate}
\title{Climate data of the Agrodiversity experiment}
\format{
A dataframe with 39,712 rows and 9 variables:
\describe{
\item{SITE}{Site ID number}
\item{DAY}{Day of data collected}
\item{MONTH}{Month of data collected}
\item{YEAR}{Year of data collected}
\item{DATE}{Date of data collection}
\item{PRECIP}{Daily precipitation (mm/day)}
\item{AIR_MIN}{Minimum daily air temperature (°C)}
\item{AIR_MEAN}{Mean daily air temperature (°C)}
\item{AIR_MAX}{Maximum daily air temperature (°C)}
}
}
\source{
Laura Kirwan, John Connolly, Caroline Brophy, Ole Baadshaug, Gilles Belanger, Alistair Black, Tim Carnus, Rosemary Collins, Jure Čop, Ignacio Delgado, Alex De Vliegher, Anjo Elgersma, Bodil Frankow-Lindberg, Piotr Golinski, Philippe Grieu, Anne-Maj Gustavsson, Áslaug Helgadóttir, Mats Höglind, Olivier Huguenin-Elie, Marit Jørgensen, Žydrė Kadžiulienė, Tor Lunnan, Andreas Lüscher, Päivi Kurki, Claudio Porqueddu, M.-Teresa Sebastia, Ulrich Thumm, David Walmsley, John Finn. 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 95(9), pp. 2680-2680.

\url{https://doi.org/10.6084/m9.figshare.3560352.v1}
}
\usage{
Climate
}
\description{
The \code{Climate} dataset has information about  daily climate data for each site for the duration of the experiment.
The climate data series begins on the sowing date of the experiment at each site.
The variables include daily temperatures and precipitation.
}
\keyword{datasets}
