% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\docType{data}
\name{Soil}
\alias{Soil}
\title{Soil data of the Agrodiversity experiment}
\format{
A dataframe with 32 rows and 17 variables:
\describe{
\item{SITE}{Site ID number}
\item{Carbonates}{Soil Carbonates (\%)}
\item{EC}{Soil electrical conductivity (ds m-1)}
\item{Silt}{Percent silt content in soil (\%)}
\item{Silt_fine}{Percent fine silt content in soil (\%)}
\item{Clay}{Percent clay content in soil (\%)}
\item{Sand}{Percent sand content in soil (\%)}
\item{OM}{Percent organic matter (\%)}
\item{Soil_Type}{Soil type}
\item{Humidity}{Percent humidity (\%)}
\item{Ca}{Calcium concentration (ppm)}
\item{K}{Potassium concentration (ppm)}
\item{N-NO3}{Nitrate concentration (ppm)}
\item{Mg}{Magnesium concentration (ppm)}
\item{Na}{Sodium concentration (ppm)}
\item{P}{Phosphorus concentration (ppm)}
\item{pH}{Soil pH}
}
}
\source{
Kirwan, L., Connolly, J., Brophy, C., Baadshaug, O.H., Belanger, G., Black, A., Carnus, T., Collins, R., Čop, J., Delgado, I., De Vliegher, A., Elgersma, A., Frankow-Lindberg, B., Golinski, P., Grieu, P., Gustavsson, A.-M., Helgadottir, A., Höglind, M., Huguenin-Elie, O., Jørgensen, M., Kadziulienė, Z., Lunnan, T., Lüscher, A., Kurki P., Porqueddu, C.,  Sebastià, M.T., Thumm, U., Walmsley, D. and Finn, J.A, 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 95(9), pp.2680-2680.

Dataset is available at: \url{https://doi.org/10.6084/m9.figshare.3560352.v1}
}
\usage{
Soil
}
\description{
The \code{Soil} dataset provides information about soil characteristics and concentration of soil nutrients for each site.
}
\keyword{datasets}
