% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather1.R
\docType{data}
\name{Weather1}
\alias{Weather1}
\title{Climate data of the Agrodiversity experiment}
\format{
A dataframe with 39,712 rows and 9 variables:
\describe{
\item{SITE}{Site ID number}
\item{DAY}{Day of data collected}
\item{MONTH}{Month of data collected}
\item{YEAR}{Year of data collected}
\item{DATE}{Date of data collection}
\item{PRECIP}{Daily precipitation (mm/day)}
\item{AIR_MIN}{Minimum daily air temperature (°C)}
\item{AIR_MEAN}{Mean daily air temperature (°C)}
\item{AIR_MAX}{Maximum daily air temperature (°C)}
}
}
\source{
Kirwan, L., Connolly, J., Brophy, C., Baadshaug, O.H., Belanger, G., Black, A., Carnus, T., Collins, R., Čop, J., Delgado, I., De Vliegher, A., Elgersma, A., Frankow-Lindberg, B., Golinski, P., Grieu, P., Gustavsson, A.-M., Helgadottir, A., Höglind, M., Huguenin-Elie, O., Jørgensen, M., Kadziulienė, Z., Lunnan, T., Lüscher, A., Kurki P., Porqueddu, C.,  Sebastià, M.T., Thumm, U., Walmsley, D. and Finn, J.A, 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 95(9), pp.2680-2680.

Dataset is available at: \url{https://doi.org/10.6084/m9.figshare.3560352.v1}
}
\usage{
Weather1
}
\description{
The \code{Weather1} dataset has information about  daily climate data for each site for the duration of the experiment.
The climate data series begins on the sowing date of the experiment at each site.
The variables include daily temperatures and precipitation.
}
\keyword{datasets}
