% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgrodiversityData-package.R
\docType{package}
\name{AgrodiversityData}
\alias{AgrodiversityData}
\title{AgrodiversityData:  Datasets of agrodiversity experiment.}
\description{
The package \code{AgrodiversityData} provides six datasets of Agrodiversity experiment.
It is  useful for ecologists studying relationships between diversity and ecosystem function and agronomists interested in sustainable intensification.
}
\section{Agrodiversity Experiment}{

A coordinated, continental-scale 33 site experiment that investigated plant diversity and ecosystem function in intensively-managed grassland communities across 17 countries.
the study investigated whether experimental increases in plant diversity in intensively managed grassland communities can increase their resource use efficiency.
Each site had 30 core plots, representing 15 grassland communities at two seeding densities.
The 15 communities were comprised of four monocultures (two grasses and two legumes) and 11 four-species mixtures that varied in the relative abundance of the four species at sowing.
Sites agreed a protocol and employed the same experimental methods with certain plot management factors, such as seeding rates and number of cuts, determined by local practice.
The four species used at a site depended on geographical location, but the species were chosen according to four functional traits: a fast-establishing grass, a slow-establishing persistent grass, a fast-establishing legume, and a slow-establishing persistent legume.
As the objective was to maximize yield for intensive grassland production, the species chosen were all high-yielding agronomic species. The data set contains species-specific biomass measurements (yield per species and of weeds) for all harvests for up to four years at 33 sites.
}

\section{Details of Datasets}{

\enumerate{
\item \code{\link{Site}}: Information about individual sites in the experiment specifying details of country, species, treatment, harvest etc. .
\item \code{\link{Climate}}: Information about daily temperature and precipitation for each site.
\item \code{\link{Soil}}: The soil characteristics and concentration of soil nutrients for each site.
\item \code{\link{Biomass}}: The species-specific biomass measurements (yield per species and of weeds) for all harvests for up to four years at 33 sites.
\item \code{\link{Forage}}: Details about forage quality.
\item \code{\link{Biosum}}: Species-specific biomass measurements (yield per species and of weeds) at 33 sites summarized by year.
}
}

\section{References}{

\enumerate{
\item Kirwan, L., Lüscher, A., Sebastià, M.T., Finn, J.A., Collins, R.P., Porqueddu, C., Helgadottir, A., Baadshaug, O.H., Brophy, C., Coran, C. and Dalmannsdottir, S., 2007. Evenness drives consistent diversity effects in intensive grassland systems across 28 European sites. Journal of Ecology, 95(3), pp.530-539.
\item Kirwan, L., Connolly, J., Finn, J.A., Brophy, C., Lüscher, A., Nyfeler, D. and Sebastià, M.T., 2009. Diversity–interaction modeling: estimating contributions of species identities and interactions to ecosystem function. Ecology, 90(8), pp.2032-2038.
\item Connolly, J., Bell, T., Bolger, T., Brophy, C., Carnus, T., Finn, J.A., Kirwan, L., Isbell, F., Levine, J., Lüscher, A. and Picasso, V., 2013. An improved model to predict the effects of changing biodiversity levels on ecosystem function. Journal of Ecology, 101(2), pp.344-355.
\item Finn, J.A., Kirwan, L., Connolly, J., Sebastià, M.T., Helgadottir, A., Baadshaug, O.H., Bélanger, G., Black, A., Brophy, C., Collins, R.P. and Čop, J., 2013. Ecosystem function enhanced by combining four functional types of plant species in intensively managed grassland mixtures: a 3‐year continental‐scale field experiment. Journal of Applied Ecology, 50(2), pp.365-375.
\item Laura Kirwan, John Connolly, Caroline Brophy, Ole Baadshaug, Gilles Belanger, Alistair Black, Tim Carnus, Rosemary Collins, Jure Čop, Ignacio Delgado, Alex De Vliegher, Anjo Elgersma, Bodil Frankow-Lindberg, Piotr Golinski, Philippe Grieu, Anne-Maj Gustavsson, Áslaug Helgadóttir, Mats Höglind, Olivier Huguenin-Elie, Marit Jørgensen, Žydrė Kadžiulienė, Tor Lunnan, Andreas Lüscher, Päivi Kurki, Claudio Porqueddu, M.-Teresa Sebastia, Ulrich Thumm, David Walmsley, John Finn. 2014. The Agrodiversity Experiment: three years of data from a multisite study in intensively managed grasslands. Ecology, 95(9), pp. 2680-2680.
\item Dooley, Á., Isbell, F., Kirwan, L., Connolly, J., Finn, J.A. and Brophy, C., 2015. Testing the effects of diversity on ecosystem multifunctionality using a multivariate model. Ecology Letters, 18(11), pp.1242-1251.
\item Brophy, C., Dooley, Á., Kirwan, L., Finn, J.A., McDonnell, J., Bell, T., Cadotte, M.W. and Connolly, J., 2017. Biodiversity and ecosystem function: making sense of numerous species interactions in multi‐species communities. Ecology, 98(7), pp.1771-1778.
\item Connolly, J., Sebastià, M.T., Kirwan, L., Finn, J.A., Llurba, R., Suter, M., Collins, R.P., Porqueddu, C., Helgadóttir, Á., Baadshaug, O.H. and Bélanger, G., 2018. Weed suppression greatly increased by plant diversity in intensively managed grasslands: A continental‐scale experiment. Journal of Applied Ecology, 55(2), pp.852-862.
}
}

\examples{

 ## Work with Biomass dataset
    library(AgrodiversityData)

 ## Load the data
    data(Biomass)
 
 ## To see the data structure
    str(Biomass)
 
 ## To get summary statistics of Total harvest Yield
    summary(Biomass$HARV_YIELD)
 
 ## To check the normality of the data
    qqnorm(Biomass$HARV_YIELD)
 
 ## To check whether mean yield is same for High and Low seed densities.
    t.test(Biomass$HARV_YIELD~Biomass$DENS)


}
\seealso{
Data source :\url{https://doi.org/10.6084/m9.figshare.3560352.v1}
}
